Score increase is slower

+++ CProcessing/src/Score/score.h
@@ -3,5 +3,5 @@
 #include "../Food/food.h"
 #include <stdlib.h>
 
-void DisplayScore(int score);
-int AddScore();
+void DisplayScore(float score);
+float AddScore();
